pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crawlers"
version = "0.1.0"
description = "Collection short videos from internet"
authors = [
    {name = "PING",email = "liping_000@126.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["crawler", "video", "tiktok", "douyin", "bilibili", "downloader"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Multimedia :: Video",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    'aiofiles==23.2.1',
    'annotated-types==0.6.0',
    'anyio==4.3.0',
    'browser-cookie3==0.19.1',
    'certifi==2024.2.2',
    'click==8.1.7',
    'colorama==0.4.6',
    'fastapi==0.110.2',
    'h11==0.14.0',
    'httpcore==1.0.5',
    'httpx==0.27.0',
    'idna==3.7',
    'importlib_resources==6.4.0',
    'lz4==4.3.3',
    'markdown-it-py==3.0.0',
    'mdurl==0.1.2',
    'numpy',
    'pycryptodomex==3.20.0',
    'pydantic==2.7.0',
    'pydantic_core==2.18.1',
    'pyfiglet==1.0.0',
    'Pygments==2.17.2',
    'pypng==0.20220715.0',
    'pywebio==1.8.3',
    'pywebio-battery==0.6.0',
    'PyYAML==6.0.1',
    'qrcode==7.4.2',
    'rich==13.7.1',
    'sniffio==1.3.1',
    'starlette==0.37.2',
    'tornado==6.4',
    'typing_extensions==4.11.0',
    'ua-parser==0.18.0',
    'user-agents==2.2.0',
    'uvicorn==0.29.0',
    'websockets==12.0',
    'gmssl==3.2.2',
    'tenacity~=9.0.0'
]

[tool.setuptools.packages.find]
where = ["."]
include = ["crawlers*"]

[tool.setuptools.package-data]
crawlers = ["**/*.yaml", "**/*.yml", "**/*.json"]

[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "*.pyo", "__pycache__", ".git*", ".idea*", "tests*", "*.test.py"]
